requirements.txt
autopep8==2.0.2
click==8.1.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
llvmlite==0.39.1
MarkupSafe==2.1.2
numba==0.56.4
numpy==1.23.5
packaging==23.1
pandas==2.0.0
plotly==5.14.1
pycodestyle==2.10.0
pynndescent==0.5.10
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
tzdata==2023.3
umap-learn==0.5.3
Werkzeug==2.2.3
